refactor(styles): split oversized CSS-in-TS files into per-component files
Splits three grab-bag *Styles.ts files into per-component/per-concern
files, co-located with their consumers, following the existing
designTokens composition idiom (no behavior change):

- requestPanelStyles.ts (1058 lines) -> requestPanelSharedStyles.ts
  (shared layout/accent/rejection-feedback machinery) plus one styles
  file per panel (ToolEdit, Bash, Retry, Proposal, ExternalInquiry,
  UserQuestion); PlanApprovalRequestPanel's ~10-line rule is now an
  inline component-local `css` literal.
- profile/styles.ts (484 lines) -> ProviderKeyList.styles.ts,
  ApiAccessSection.styles.ts, ModelSelectionList.styles.ts, one per
  consumer; also drops a dead `profileViewStyles` import from
  RelayQuotaMeter.ts (it only ever used its own inline styles).
- streamTabsStyles.ts (428 lines) -> StreamTab.styles.ts +
  StreamTabsContainer.styles.ts, matching the two custom elements
  StreamTabs.ts already defines.
