+++ packages/extension/src/progressView/frontend/components/StreamTabs.ts
@@ -40,7 +40,8 @@ import { formatRelativeTime } from '@shared/utils/string';
 import { renderIconActionButton } from '@shared/wa/actionButtons';
 import { TEXRA_ICON_LIBRARY, waIcon } from '@shared/wa/webAwesomeIcons';
 import { layoutStyles } from '../styles/logStyles';
-import { streamTabStyles, streamTabsContainerStyles } from './streamTabsStyles';
+import { streamTabStyles } from './StreamTab.styles';
+import { streamTabsContainerStyles } from './StreamTabsContainer.styles';
 import { ELEMENT_IDS, FILTER_BUTTONS } from '../constants';
 import { ProgressEvents } from '../events';
 import { getComposedPathElement, getRadioValue, setsEqual } from '../utils';
