--- src/shared/styles/requestPanelStyles.ts
@@ -1,1041 +1,0 @@
-import { css, unsafeCSS, type CSSResult } from 'lit';
-
-/**
- * Single source of truth for request panel types. Each entry pairs the outer
- * container class (plural), the item class (singular, used for cards and BEM
- * children), and the accent color used for the item's left border and its
- * matching header icon. Add a panel type here and the shared layout, accent,
- * and icon rules all follow — no parallel arrays to keep in sync.
- */
-const PANEL_TYPES = [
-  {
-    container: 'approval-requests',
-    item: 'approval-request',
-    accent: 'var(--wa-color-text-normal)',
-  },
-  {
-    container: 'bash-approval-requests',
-    item: 'bash-approval-request',
-    accent: 'var(--wa-color-terminal-ansi-yellow)',
-  },
-  {
-    container: 'retry-requests',
-    item: 'retry-request',
-    accent: 'var(--color-warning)',
-  },
-  {
-    container: 'workflow-proposals',
-    item: 'workflow-proposal',
-    accent: 'var(--wa-color-text-link)',
-  },
-  {
-    container: 'plan-approval-requests',
-    item: 'plan-approval-request',
-    accent: 'var(--wa-color-text-link)',
-  },
-  {
-    container: 'external-inquiry-requests',
-    item: 'external-inquiry-request',
-    accent: 'var(--wa-color-focus)',
-  },
-  {
-    container: 'user-question-requests',
-    item: 'user-question-request',
-    accent: 'var(--wa-color-focus)',
-  },
-] as const;
-
-/** Container class names (plural, used as the outer wrapper). */
-const CONTAINER_NAMES = PANEL_TYPES.map((p) => p.container);
-
-/** Item class names (singular, used for individual cards and BEM children). */
-const ITEM_NAMES = PANEL_TYPES.map((p) => p.item);
-
-/** Build a :is()-ready selector group from class names with an optional BEM suffix. */
-function selectorGroup(names: readonly string[], suffix = ''): CSSResult {
-  return unsafeCSS(names.map((n) => `.${n}${suffix}`).join(', '));
-}
-
-const CONTAINERS = selectorGroup(CONTAINER_NAMES);
-const HEADERS = selectorGroup(CONTAINER_NAMES, '__header');
-const LISTS = selectorGroup(CONTAINER_NAMES, '__list');
-const ITEMS = selectorGroup(ITEM_NAMES);
-const DETAILS = selectorGroup(ITEM_NAMES, '__details');
-const ACTIONS = selectorGroup(ITEM_NAMES, '__actions');
-const FEEDBACKS = selectorGroup(ITEM_NAMES, '__feedback');
-const FEEDBACK_INPUTS = selectorGroup(ITEM_NAMES, '__feedback-input');
-const FEEDBACK_ACTIVE = selectorGroup(ITEM_NAMES, '--feedback-active');
-const SCROLLABLE_DETAILS = selectorGroup(
-  ITEM_NAMES.filter((n) => n !== 'workflow-proposal'),
-  '__details',
-);
-
-/** Per-type accent rules: item left-border plus matching header icon color. */
-const ACCENT_RULES = unsafeCSS(
-  PANEL_TYPES.map(
-    ({ item, container, accent }) =>
-      `.${item} {
-    border-left: var(--border-medium) solid ${accent};
-  }
-  .${container}__header wa-icon {
-    color: ${accent};
-  }`,
-  ).join('\n  '),
-);
-
-const sp = {
-  tiny: unsafeCSS('var(--wa-space-3xs)'),
-  small: unsafeCSS('var(--wa-space-2xs)'),
-  medium: unsafeCSS('var(--wa-space-xs)'),
-  large: unsafeCSS('var(--wa-space-s)'),
-  xlarge: unsafeCSS('var(--wa-space-l)'),
-} as const;
-
-export const requestPanelStyles: CSSResult = css`
-  /* ================================================================
-   * Shared request panel layout (all panel types)
-   * ================================================================ */
-
-  :host {
-    box-sizing: border-box;
-    min-width: 0;
-    max-width: 100%;
-  }
-
-  :is(${CONTAINERS}) {
-    margin: ${sp.medium} 0;
-    padding: ${sp.medium};
-    border: var(--border-thin) solid var(--wa-form-control-border-color);
-    border-radius: var(--border-radius-large);
-    background: var(--wa-color-surface-default);
-    display: flex;
-    flex-direction: column;
-    gap: ${sp.medium};
-    box-sizing: border-box;
-    min-width: 0;
-    max-width: 100%;
-  }
-
-  :is(${HEADERS}) {
-    display: flex;
-    align-items: center;
-    gap: ${sp.medium};
-    font-weight: var(--font-weight-semibold);
-    color: var(--wa-color-text-normal);
-  }
-
-  :is(${LISTS}) {
-    display: flex;
-    flex-direction: column;
-    gap: ${sp.medium};
-    min-width: 0;
-    max-width: 100%;
-  }
-
-  :is(${ITEMS}) {
-    display: flex;
-    flex-direction: column;
-    border-radius: var(--border-radius);
-    border: var(--border-thin) solid var(--wa-color-surface-border);
-    background: var(--wa-color-surface-raised);
-    padding: ${sp.medium};
-    gap: ${sp.medium};
-    position: relative;
-    box-sizing: border-box;
-    min-width: 0;
-    max-width: 100%;
-  }
-
-  :is(${DETAILS}) {
-    display: flex;
-    flex-direction: column;
-    gap: ${sp.small};
-  }
-
-  /* Scroll constraint for panels with potentially long content.
-     Excluded: workflow-proposal__details (has an absolutely-positioned
-     dropdown that would be clipped by overflow-y: auto). */
-  :is(${SCROLLABLE_DETAILS}) {
-    max-height: min(34vh, 26rem);
-    overflow-y: auto;
-    scrollbar-gutter: stable;
-  }
-
-  .approval-request__meta,
-  .retry-request__meta {
-    font-size: var(--font-size-sm);
-    color: var(--color-text-secondary);
-    display: flex;
-    align-items: center;
-    gap: ${sp.small};
-    flex-wrap: wrap;
-  }
-
-  :is(${ACTIONS}) {
-    display: flex;
-    align-items: center;
-    flex-wrap: wrap;
-    gap: ${sp.small};
-    min-width: 0;
-    max-width: 100%;
-  }
-
-  :is(${ACTIONS}) > * {
-    min-width: 0;
-    max-width: 100%;
-  }
-
-  /* The primary action buttons hug their content and stay button-sized: they
-     never grow to fill a wide panel and are width-capped, so they group at the
-     start of the row instead of stretching, and the label centers within the
-     button's own padding. min-width: auto restores the flex min-content floor
-     (overriding the bounded-children min-width: 0 above), so a button never
-     shrinks below its own label and wraps to the next row instead of
-     clipping. Keyed off the semantic .action-button class the button helper
-     sets in its template (not the data-action behaviour hook) and sized on the
-     host element, so there is no shadow-part piercing and no source-order
-     tiebreak to keep in sync. Tool edit approvals override with their bespoke
-     sizing below, winning by specificity. */
-  :is(${ACTIONS}) .action-button {
-    flex: 0 1 auto;
-    min-width: auto;
-    max-width: min(14rem, 100%);
-  }
-
-  .approval-request__actions wa-button[data-action] {
-    flex: 0 1 7rem;
-    width: auto;
-    min-width: min(5.5rem, 100%);
-    max-width: min(7rem, 100%);
-  }
-
-  .approval-request__actions wa-button[data-action]::part(base) {
-    justify-content: center;
-    width: 100%;
-    inline-size: 100%;
-    min-width: 0;
-    max-width: 100%;
-  }
-
-  .approval-request__actions wa-button[data-action]::part(label),
-  .approval-request__actions .diff-dropdown .diff-main-button::part(label) {
-    min-width: 0;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-  }
-
-  /* Shared action button colors */
-  :is(${ACTIONS}) wa-button[data-action='approve']::part(base),
-  :is(${ACTIONS}) wa-button[data-action='submit']::part(base) {
-    color: var(--wa-color-success-fill-loud);
-  }
-
-  :is(${ACTIONS}) wa-button[data-action='setup']::part(base) {
-    color: var(--wa-color-text-link);
-  }
-
-  :is(${FEEDBACK_ACTIVE})
-    :is(${ACTIONS})
-    wa-button[data-action='reject']::part(base) {
-    color: var(--wa-color-warning-border-quiet);
-  }
-
-  /* Type-specific item accent + matching header icon color (see PANEL_TYPES / ACCENT_RULES) */
-  ${ACCENT_RULES}
-
-  /* ================================================================
-   * Approval requests (tool edits)
-   * ================================================================ */
-
-  .approval-request__path {
-    font-family: var(--wa-font-family-mono);
-    font-size: var(--font-size-sm);
-    color: var(--color-text-link);
-    word-break: break-word;
-  }
-
-  .approval-request__diff {
-    display: inline-flex;
-    align-items: baseline;
-    gap: ${sp.small};
-  }
-
-  .approval-request__diff-added,
-  .approval-request__diff-removed {
-    font-size: var(--font-size-xs);
-  }
-
-  .approval-request__diff-added {
-    color: var(--color-added);
-  }
-
-  .approval-request__diff-removed {
-    color: var(--color-removed);
-  }
-
-  .approval-request__diff-label {
-    color: var(--color-text-secondary);
-    font-size: var(--font-size-xs);
-    opacity: var(--opacity-normal);
-  }
-
-  .approval-request__actions .diff-dropdown {
-    position: relative;
-    display: inline-flex;
-    align-items: stretch;
-    flex: 1 1 7rem;
-    min-width: 0;
-    max-width: min(8.25rem, 100%);
-  }
-
-  .approval-request__actions .diff-dropdown .diff-main-button {
-    flex: 1 1 auto;
-    width: auto;
-    min-width: 0;
-  }
-
-  /* Square the inner corners so the label and caret fuse into one pill. */
-  .approval-request__actions .diff-dropdown .diff-main-button::part(base) {
-    width: 100%;
-    min-width: 0;
-    justify-content: center;
-    border-top-right-radius: 0;
-    border-bottom-right-radius: 0;
-  }
-
-  /* wa-dropdown handles its own popup positioning; pull the caret onto the
-     label so their 1px borders overlap into a single divider on hover. */
-  .approval-request__actions .diff-dropdown .diff-dropdown-menu {
-    flex: 0 0 auto;
-    display: inline-flex;
-    margin-left: calc(-1 * var(--border-thin));
-  }
-
-  .approval-request__actions .diff-dropdown .diff-dropdown-trigger {
-    flex: 0 0 auto;
-    width: 1.5rem;
-    min-width: 1.5rem;
-  }
-
-  /* Match the .action-button chrome: borderless at rest with the border
-     reserved (transparent) so nothing shifts, caret at full presence, and only
-     the right corners rounded so it tucks against the label. */
-  .approval-request__actions .diff-dropdown .diff-dropdown-trigger::part(base) {
-    min-height: var(--height-control-compact);
-    height: auto;
-    width: 100%;
-    padding: 0;
-    opacity: var(--opacity-full);
-    background: transparent;
-    border: var(--border-thin) solid transparent;
-    border-top-left-radius: 0;
-    border-bottom-left-radius: 0;
-  }
-
-  .approval-request__actions
-    .diff-dropdown
-    .diff-dropdown-trigger::part(base):hover {
-    background: transparent;
-  }
-
-  .approval-request__actions .diff-dropdown .diff-dropdown-trigger wa-icon {
-    font-size: var(--font-size-sm);
-  }
-
-  /* Hovering or opening either half outlines the whole pair as one box with a
-     single internal divider, so the caret reads as the diff button's menu
-     rather than a stray glyph floating beside it. */
-  .approval-request__actions .diff-dropdown:hover .diff-main-button::part(base),
-  .approval-request__actions
-    .diff-dropdown:hover
-    .diff-dropdown-trigger::part(base),
-  .approval-request__actions
-    .diff-dropdown:focus-within
-    .diff-main-button::part(base),
-  .approval-request__actions
-    .diff-dropdown:focus-within
-    .diff-dropdown-trigger::part(base),
-  .approval-request__actions
-    .diff-dropdown
-    wa-dropdown[open]
-    .diff-main-button::part(base),
-  .approval-request__actions
-    .diff-dropdown
-    wa-dropdown[open]
-    .diff-dropdown-trigger::part(base) {
-    border-color: var(--wa-color-surface-border, var(--color-border));
-  }
-
-  .approval-request__actions
-    .diff-dropdown
-    .diff-dropdown-trigger:focus-visible::part(base) {
-    outline: var(--border-thin) solid var(--wa-color-focus);
-    outline-offset: var(--border-thin);
-  }
-
-  .approval-request__actions
-    .diff-dropdown
-    wa-dropdown[open]
-    .diff-dropdown-trigger
-    wa-icon {
-    transform: rotate(180deg);
-  }
-
-  /* ================================================================
-   * Bash approval requests
-   * ================================================================ */
-
-  .bash-approval-request__command {
-    font-family: var(--wa-font-family-mono);
-    font-size: var(--font-size-sm);
-  }
-
-  .bash-approval-request__cwd {
-    color: var(--wa-color-text-quiet);
-    font-family: var(--wa-font-family-mono);
-    font-size: var(--font-size-xs);
-    margin-bottom: var(--wa-space-xs);
-    overflow-wrap: anywhere;
-  }
-
-  .bash-approval-request__cwd span {
-    color: var(--color-text-secondary);
-  }
-
-  .bash-approval-request__command .code-block {
-    border-left: var(--border-medium) solid var(--wa-color-terminal-ansi-yellow);
-  }
-
-  .bash-approval-request__command .code-block pre {
-    white-space: pre-wrap;
-    word-break: break-word;
-  }
-
-  /* ================================================================
-   * Retry requests
-   * ================================================================ */
-
-  .retry-request__operation {
-    font-family: var(--wa-font-family-mono);
-    font-size: var(--font-size-sm);
-    color: var(--wa-color-text-normal);
-    font-weight: var(--font-weight-medium);
-  }
-
-  .retry-request__error {
-    font-size: var(--font-size-xs);
-    color: var(--color-error);
-    word-break: break-word;
-    max-height: 4em;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-
-  .retry-request__error-details {
-    margin-top: ${sp.tiny};
-    font-size: var(--font-size-xs);
-  }
-
-  .retry-request__error-details::part(header) {
-    min-height: 28px;
-    padding: ${sp.small} ${sp.large};
-    color: var(--wa-color-text-quiet);
-  }
-
-  .retry-request__error-details::part(content) {
-    padding: 0 ${sp.large} ${sp.large};
-  }
-
-  .retry-request__error-summary {
-    cursor: pointer;
-    color: var(--wa-color-text-quiet);
-    display: flex;
-    align-items: center;
-    gap: ${sp.tiny};
-    user-select: none;
-  }
-
-  .retry-request__error-summary:hover {
-    color: var(--wa-color-text-normal);
-  }
-
-  /* Note: .toggle-icon rotation handled by commonViewStyles */
-
-  .retry-request__error-body {
-    margin-top: ${sp.tiny};
-    padding: ${sp.small};
-    background: var(--wa-color-surface-lowered, rgba(0, 0, 0, 0.1));
-    border-radius: var(--border-radius-small);
-    font-family: var(--wa-font-family-mono);
-    font-size: var(--font-size-xs);
-    white-space: pre-wrap;
-    word-break: break-word;
-    max-height: 12em;
-    overflow-y: auto;
-  }
-
-  .retry-request--relay .retry-request__operation {
-    color: var(--wa-color-warning-on-quiet, #ff8c00);
-  }
-
-  /* ================================================================
-   * Workflow proposals
-   * ================================================================ */
-
-  .workflow-proposal__header-row {
-    display: flex;
-    align-items: baseline;
-    gap: ${sp.medium};
-    flex-wrap: wrap;
-  }
-
-  .workflow-proposal__agent {
-    font-family: var(--wa-font-family-mono);
-    font-size: var(--font-size);
-    font-weight: var(--font-weight-semibold);
-    color: var(--wa-color-text-link);
-  }
-
-  .workflow-proposal__model {
-    font-size: var(--font-size-sm);
-    color: var(--color-text-secondary);
-  }
-
-  .workflow-proposal__model::before {
-    content: '\u2022';
-    margin-right: ${sp.small};
-  }
-
-  .workflow-proposal__agent-select,
-  .workflow-proposal__model-select {
-    display: flex;
-    align-items: center;
-    gap: ${sp.small};
-  }
-
-  .workflow-proposal__agent-select wa-icon,
-  .workflow-proposal__model-select wa-icon {
-    color: var(--wa-color-text-quiet);
-    flex-shrink: 0;
-  }
-
-  /* Model dropdown floats to the right edge of the header row. */
-  .workflow-proposal__model-select {
-    margin-left: auto;
-  }
-
-  .proposal-model-dropdown {
-    min-width: 6rem;
-    max-width: 9rem;
-  }
-
-  .proposal-agent-dropdown {
-    min-width: 5rem;
-    max-width: 8rem;
-  }
-
-  .workflow-proposal__instruction {
-    font-size: var(--font-size-sm);
-    color: var(--wa-color-text-normal);
-    word-break: break-word;
-    max-height: 12em;
-    overflow-y: auto;
-    line-height: var(--line-height-normal);
-    padding: ${sp.small} 0;
-    border-bottom: var(--border-thin) solid var(--wa-color-surface-border);
-  }
-
-  .workflow-proposal__instruction pre {
-    white-space: pre-wrap;
-  }
-
-  .workflow-proposal__extract-flags {
-    display: flex;
-    gap: ${sp.small};
-    flex-wrap: wrap;
-    margin-top: ${sp.small};
-  }
-
-  .workflow-proposal__files {
-    display: flex;
-    flex-direction: column;
-    gap: ${sp.small};
-    margin-top: ${sp.small};
-  }
-
-  .workflow-proposal__files > div {
-    font-size: var(--font-size-sm);
-    color: var(--wa-color-text-normal);
-    line-height: var(--line-height-normal);
-  }
-
-  .workflow-proposal__file-label {
-    font-weight: var(--font-weight-semibold);
-    color: var(--color-text-secondary);
-  }
-
-  .workflow-proposal__file-name {
-    font-family: var(--wa-font-family-mono);
-    color: var(--wa-color-text-link);
-    cursor: pointer;
-  }
-
-  .workflow-proposal__file-name:hover {
-    text-decoration: underline;
-    color: var(--wa-color-text-link-active, var(--wa-color-text-link));
-  }
-
-  .workflow-proposal__file-name--readonly {
-    color: var(--color-text-secondary);
-    cursor: default;
-  }
-
-  .workflow-proposal__file-name--readonly:hover {
-    text-decoration: none;
-    color: var(--color-text-secondary);
-  }
-
-  .workflow-proposal__file-name--wrap {
-    word-break: break-word;
-  }
-
-  .workflow-proposal__input-files .workflow-proposal__file-label {
-    color: var(--wa-color-text-normal);
-  }
-
-  .workflow-proposal__output-files .workflow-proposal__file-label {
-    color: var(--wa-color-text-link);
-  }
-
-  /* ================================================================
-   * Plan approval requests
-   * ================================================================ */
-
-  .plan-approval-request__objective {
-    margin: ${sp.small} 0;
-    color: var(--wa-color-text-normal);
-    white-space: pre-wrap;
-    word-break: break-word;
-  }
-
-  /* ================================================================
-   * Rejection feedback (shared across panels with feedback support)
-   * ================================================================ */
-
-  :is(${FEEDBACKS}) {
-    margin-top: ${sp.small};
-  }
-
-  :is(${FEEDBACK_INPUTS}) {
-    width: 100%;
-  }
-
-  /* ================================================================
-   * External inquiry requests
-   * ================================================================ */
-
-  /* Carousel navigation for multiple external inquiries */
-  .external-inquiry-requests__nav {
-    display: flex;
-    align-items: center;
-    gap: ${sp.small};
-    margin-left: auto;
-  }
-
-  .external-inquiry-requests__counter {
-    font-size: var(--font-size-sm);
-    color: var(--wa-color-text-quiet);
-    font-variant-numeric: tabular-nums;
-    min-width: 3ch;
-    text-align: center;
-  }
-
-  .external-inquiry-request__context {
-    font-size: var(--font-size-sm);
-    color: var(--wa-color-text-quiet);
-    padding: ${sp.small} ${sp.medium};
-    background: var(--wa-color-surface-lowered, rgba(0, 0, 0, 0.1));
-    border-radius: var(--border-radius-small);
-    line-height: var(--line-height-normal);
-  }
-
-  .external-inquiry-request__question {
-    background: var(--wa-color-surface-lowered, rgba(0, 0, 0, 0.05));
-    border: var(--border-thin) solid var(--wa-color-surface-border);
-    border-radius: var(--border-radius);
-    padding: ${sp.medium};
-    position: relative;
-  }
-
-  .external-inquiry-request__question-text {
-    font-size: var(--font-size);
-    line-height: var(--line-height-relaxed, 1.6);
-    color: var(--wa-color-text-normal);
-    white-space: pre-wrap;
-    word-break: break-word;
-    max-height: min(22vh, 14rem);
-    overflow-y: auto;
-    scrollbar-gutter: stable;
-  }
-
-  .external-inquiry-request__question-actions {
-    display: flex;
-    justify-content: flex-end;
-    margin-top: ${sp.small};
-    padding-top: ${sp.small};
-    border-top: var(--border-thin) solid var(--wa-color-surface-border);
-  }
-
-  .external-inquiry-request__transcript {
-    border: var(--border-thin) solid var(--wa-color-surface-border);
-    border-radius: var(--border-radius);
-    background: var(--wa-color-surface-lowered, rgba(0, 0, 0, 0.05));
-  }
-
-  .external-inquiry-request__transcript::part(base) {
-    border: 0;
-    border-radius: 0;
-    background: transparent;
-  }
-
-  .external-inquiry-request__transcript::part(header),
-  .external-inquiry-request__transcript::part(content) {
-    padding: 0;
-  }
-
-  /* Web Awesome's details template exposes the disclosure indicator as part="icon". */
-  .external-inquiry-request__transcript::part(icon) {
-    padding-inline-end: ${sp.medium};
-  }
-
-  .external-inquiry-request__transcript-summary {
-    display: flex;
-    align-items: center;
-    gap: ${sp.small};
-    padding: ${sp.small} ${sp.medium};
-    cursor: pointer;
-    color: var(--wa-color-text-normal);
-    font-size: var(--font-size-sm);
-    font-weight: var(--font-weight-semibold);
-  }
-
-  .external-inquiry-request__transcript-turns {
-    display: flex;
-    flex-direction: column;
-    gap: ${sp.medium};
-    max-height: min(28vh, 18rem);
-    overflow-y: auto;
-    padding: ${sp.medium};
-    border-top: var(--border-thin) solid var(--wa-color-surface-border);
-    scrollbar-gutter: stable;
-  }
-
-  .external-inquiry-request__transcript-turn {
-    display: flex;
-    flex-direction: column;
-    gap: ${sp.small};
-  }
-
-  .external-inquiry-request__transcript-turn
-    + .external-inquiry-request__transcript-turn {
-    padding-top: ${sp.medium};
-    border-top: var(--border-thin) solid var(--wa-color-surface-border);
-  }
-
-  .external-inquiry-request__transcript-turn-header,
-  .external-inquiry-request__transcript-label {
-    font-size: var(--font-size-sm);
-    font-weight: var(--font-weight-semibold);
-    color: var(--wa-color-text-quiet);
-  }
-
-  .external-inquiry-request__transcript-context {
-    font-size: var(--font-size-sm);
-    color: var(--wa-color-text-quiet);
-  }
-
-  .external-inquiry-request__transcript-text {
-    white-space: pre-wrap;
-    word-break: break-word;
-    color: var(--wa-color-text-normal);
-    font-size: var(--font-size-sm);
-    line-height: var(--line-height-normal);
-  }
-
-  .external-inquiry-request__transcript-links {
-    display: flex;
-    flex-direction: column;
-    gap: ${sp.tiny};
-  }
-
-  .external-inquiry-request__search-hint {
-    display: flex;
-    align-items: center;
-    gap: ${sp.small};
-    font-size: var(--font-size-sm);
-    color: var(--wa-color-brand-on-quiet, var(--wa-color-focus));
-    padding: ${sp.small} 0;
-  }
-
-  .external-inquiry-request__attach-files {
-    display: flex;
-    flex-direction: column;
-    gap: ${sp.small};
-  }
-
-  .external-inquiry-request__attach-label {
-    font-size: var(--font-size-sm);
-    font-weight: var(--font-weight-semibold);
-    color: var(--wa-color-text-quiet);
-    display: flex;
-    align-items: center;
-    gap: ${sp.small};
-  }
-
-  .external-inquiry-request__file-list {
-    display: flex;
-    flex-direction: column;
-    gap: ${sp.tiny};
-    padding: ${sp.small};
-    background: var(--wa-color-surface-lowered, rgba(0, 0, 0, 0.1));
-    border-radius: var(--border-radius-small);
-  }
-
-  .external-inquiry-request__file-item {
-    display: flex;
-    align-items: center;
-    gap: ${sp.small};
-    font-size: var(--font-size-sm);
-    font-family: var(--wa-font-family-mono);
-    color: var(--wa-color-text-link);
-  }
-
-  .external-inquiry-request__answer-area {
-    display: flex;
-    flex-direction: column;
-    gap: ${sp.small};
-  }
-
-  .external-inquiry-request__session-links {
-    display: flex;
-    flex-direction: column;
-    gap: ${sp.small};
-  }
-
-  .external-inquiry-request__session-links-known,
-  .external-inquiry-request__session-links-input-group {
-    display: flex;
-    flex-direction: column;
-    gap: ${sp.small};
-  }
-
-  .external-inquiry-request__session-links-label {
-    font-size: var(--font-size-sm);
-    font-weight: var(--font-weight-semibold);
-    color: var(--wa-color-text-normal);
-  }
-
-  .external-inquiry-request__session-links-list {
-    display: flex;
-    flex-direction: column;
-    gap: ${sp.tiny};
-    padding: ${sp.small};
-    background: var(--wa-color-surface-lowered, rgba(0, 0, 0, 0.1));
-    border-radius: var(--border-radius-small);
-  }
-
-  .external-inquiry-request__session-link-item {
-    display: block;
-    font-size: var(--font-size-sm);
-    font-family: var(--wa-font-family-mono);
-    color: var(--wa-color-text-link);
-    text-decoration: none;
-    word-break: break-word;
-  }
-
-  a.external-inquiry-request__session-link-item:hover {
-    text-decoration: underline;
-  }
-
-  .external-inquiry-request__session-links-input {
-    width: 100%;
-    min-height: 2.75rem;
-    max-height: min(12vh, 5rem);
-    resize: vertical;
-    font-family: var(--wa-font-family-mono);
-    font-size: var(--font-size-sm);
-    line-height: var(--line-height-normal);
-    padding: ${sp.small} ${sp.medium};
-    background: var(--wa-form-control-background-color);
-    color: var(--wa-form-control-text-color);
-    border: var(--border-thin) solid var(--wa-form-control-border-color);
-    border-radius: var(--border-radius);
-    outline: none;
-    transition: border-color var(--transition-fast);
-  }
-
-  .external-inquiry-request__session-links-input:focus {
-    border-color: var(--wa-color-focus);
-  }
-
-  .external-inquiry-request__session-links-input::placeholder {
-    color: var(--wa-color-text-placeholder);
-  }
-
-  .external-inquiry-request__session-links-hint {
-    font-size: var(--font-size-sm);
-    color: var(--wa-color-text-quiet);
-  }
-
-  .external-inquiry-request__chat-links {
-    font-size: var(--font-size-sm);
-    color: var(--wa-color-text-quiet);
-  }
-
-  .external-inquiry-request__chat-links a {
-    color: var(--wa-color-text-link);
-    text-decoration: none;
-  }
-
-  .external-inquiry-request__chat-links a:hover {
-    text-decoration: underline;
-  }
-
-  .external-inquiry-request__answer-label {
-    font-size: var(--font-size-sm);
-    font-weight: var(--font-weight-semibold);
-    color: var(--wa-color-text-normal);
-  }
-
-  .external-inquiry-request__answer-input {
-    width: 100%;
-    min-height: 96px;
-    max-height: min(24vh, 12rem);
-    resize: vertical;
-    font-family: var(--wa-font-family-mono);
-    font-size: var(--font-size);
-    line-height: var(--line-height-normal);
-    padding: ${sp.medium};
-    background: var(--wa-form-control-background-color);
-    color: var(--wa-form-control-text-color);
-    border: var(--border-thin) solid var(--wa-form-control-border-color);
-    border-radius: var(--border-radius);
-    outline: none;
-    transition: border-color var(--transition-fast);
-  }
-
-  .external-inquiry-request__answer-input:focus {
-    border-color: var(--wa-color-focus);
-  }
-
-  .external-inquiry-request__answer-input::placeholder {
-    color: var(--wa-color-text-placeholder);
-  }
-
-  .external-inquiry-request__answer-hint {
-    font-size: var(--font-size-sm);
-    color: var(--wa-color-text-quiet);
-  }
-
-  @media (max-height: 900px) {
-    :is(${SCROLLABLE_DETAILS}) {
-      max-height: min(28vh, 20rem);
-    }
-
-    .external-inquiry-request__question-text {
-      max-height: min(18vh, 10rem);
-    }
-
-    .external-inquiry-request__answer-input {
-      min-height: 80px;
-      max-height: min(20vh, 10rem);
-    }
-
-    .external-inquiry-request__session-links-input {
-      min-height: 2.5rem;
-      max-height: min(10vh, 4.5rem);
-    }
-  }
-
-  /* ================================================================
-   * User question requests
-   * ================================================================ */
-
-  .user-question-request__context {
-    font-size: var(--font-size-sm);
-    color: var(--wa-color-text-quiet);
-    padding: ${sp.small} ${sp.medium};
-    background: var(--wa-color-surface-lowered, rgba(0, 0, 0, 0.1));
-    border-radius: var(--border-radius-small);
-    line-height: var(--line-height-normal);
-  }
-
-  .user-question-request__questions {
-    display: flex;
-    flex-direction: column;
-    gap: ${sp.medium};
-  }
-
-  .user-question-request__question {
-    display: flex;
-    flex-direction: column;
-    gap: ${sp.small};
-    padding: ${sp.medium};
-    border: var(--border-thin) solid var(--wa-color-surface-border);
-    border-radius: var(--border-radius);
-    background: var(--wa-color-surface-lowered, rgba(0, 0, 0, 0.05));
-  }
-
-  .user-question-request__heading {
-    display: flex;
-    align-items: baseline;
-    gap: ${sp.small};
-    color: var(--wa-color-text-normal);
-    font-weight: var(--font-weight-semibold);
-    line-height: var(--line-height-normal);
-  }
-
-  .user-question-request__header {
-    color: var(--wa-color-text-quiet);
-    font-size: var(--font-size-sm);
-    font-weight: var(--font-weight-semibold);
-    flex: 0 0 auto;
-  }
-
-  .user-question-request__options {
-    display: flex;
-    flex-direction: column;
-    gap: ${sp.small};
-  }
-
-  .user-question-request__option {
-    display: grid;
-    grid-template-columns: auto 1fr;
-    gap: ${sp.small};
-    align-items: start;
-    font-size: var(--font-size-sm);
-    line-height: var(--line-height-normal);
-    color: var(--wa-color-text-normal);
-    cursor: pointer;
-  }
-
-  .user-question-request__option input {
-    margin-top: 0.2rem;
-  }
-
-  .user-question-request__option span {
-    display: flex;
-    flex-direction: column;
-    gap: ${sp.tiny};
-  }
-
-  .user-question-request__option small {
-    color: var(--wa-color-text-quiet);
-    font-size: var(--font-size-xs);
-  }
-
-  .user-question-request__free-text {
-    width: 100%;
-  }
-`;
